improved cooldown notice and reduction

+++ src/main/java/de/raidcraft/skills/api/ability/Ability.java
@@ -56,7 +56,7 @@ public interface Ability<T extends CharacterTemplate> {
 
     void setRemainingCooldown(double cooldown);
 
-    long getRemainingCooldown();
+    double getRemainingCooldown();
 
     boolean isOnCooldown();
